510-scene-cleanup-shutdown-idempotency: Run Scene cleanup on SHUTDOWN as well as DESTROY and make it idempotent

+++ phaser-plus/src/engine/Scene.ts
@@ -25,6 +25,8 @@ export default class Scene extends PhaserScene {
 
     private layerSortFlag: boolean = false
 
+    private destroyed: boolean = false
+
     /** @protected */
     onInit(): void {}
 
@@ -73,7 +75,9 @@ export default class Scene extends PhaserScene {
     protected beforeInit(): void {}
 
     init(): void {
+        this.destroyed = false
         this.engine = this.initializeEngine()
+        this.events.once(Scenes.Events.SHUTDOWN, this.doDestroy, this)
         this.events.once(Scenes.Events.DESTROY, this.doDestroy, this)
         this.logger = this.engine.getLogger(`scene=${this.scene.key}`)
         this.services = this.engine.services
@@ -122,6 +126,10 @@ export default class Scene extends PhaserScene {
     }
 
     private doDestroy(): void {
+        if (this.destroyed) return
+        this.destroyed = true
+        this.events.off(Scenes.Events.SHUTDOWN, this.doDestroy, this)
+        this.events.off(Scenes.Events.DESTROY, this.doDestroy, this)
         this.features.off(LAYER_DEPTH_UPDATE, this.onLayerDepthUpdate, this)
         this.onDestroy()
         this.features.destroyAll()
